Fixed bug in parallel runs where atoms/points on the border between two or more processors were not counted.  Assigns weights to pvec/atom->partID.  Creates global grid points/spacing, and assigns these values to each partition appropriately.

+++ src/generic/vatom.c
@@ -55,7 +55,7 @@ VPUBLIC double *Vatom_getPosition(Vatom *thee) {
 
 }
 
-VPUBLIC int Vatom_getPartID(Vatom *thee) {
+VPUBLIC double Vatom_getPartID(Vatom *thee) {
 
    VASSERT(thee != VNULL);
    return thee->partID;
@@ -65,7 +65,7 @@ VPUBLIC int Vatom_getPartID(Vatom *thee) {
 VPUBLIC void Vatom_setPartID(Vatom *thee, int partID) {
 
    VASSERT(thee != VNULL);
-   thee->partID = partID;
+   thee->partID = (double)partID;
 
 }
